Refactor:큐레이션 상세 페이지 아토믹 디자인으로 구조 정리
