+++ app/feature/curation/components/CurationScrap/CurationScrapIcon.tsx
